Refactor and optimize Game of Life implementation
Major improvements:
- Performance: Use singleton CellState instances (ALIVE/DEAD) to prevent memory allocation
- API clarity: Rename Cell.translate() to withOffset() for better semantic meaning
- Type safety: Change liveNeighbors from long to int (max value is 8)
- Testing: Add assertAll groupings for better test diagnostics
- Best practices: Replace Optional.get() with orElseThrow() and proper validation
- Modern Java: Use underscore for unused lambda parameters (Java 22+ feature)

Project updates:
- Rename project from gameoflife-java21 to gameoflife
- Add MIT LICENSE for open source publication
- Add CLAUDE.md with project context and guidelines
- Update README with current versions and structure

All 42 tests pass with Java 25, Gradle 9.0, and JUnit 6.0.0-RC3.

+++ src/main/java/com/example/gameoflife/CellState.java
@@ -13,7 +13,7 @@ public sealed interface CellState permits CellState.Alive, CellState.Dead {
         public char symbol() { return '█'; }
 
         @Override
-        public CellState next(long liveNeighbors) {
+        public CellState next(int liveNeighbors) {
             return (liveNeighbors == 2 || liveNeighbors == 3) ? this : DEAD;
         }
     }
@@ -26,14 +26,14 @@ public sealed interface CellState permits CellState.Alive, CellState.Dead {
         public char symbol() { return '░'; }
 
         @Override
-        public CellState next(long liveNeighbors) {
+        public CellState next(int liveNeighbors) {
             return liveNeighbors == 3 ? ALIVE : this;
         }
     }
 
     boolean isAlive();
     char symbol();
-    CellState next(long liveNeighbors);
+    CellState next(int liveNeighbors);
 
     static CellState fromBoolean(boolean alive) {
         return alive ? ALIVE : DEAD;
